+++ src/main/java/com/arcanc/pulselib/data/gecko/PGeckoModelLoader.java
@@ -13,6 +13,7 @@ package com.arcanc.pulselib.data.gecko;
 import com.arcanc.pulselib.content.model.PModel;
 import com.arcanc.pulselib.data.PModelLoader;
 import com.arcanc.pulselib.util.PLibDatabase;
+import com.mojang.blaze3d.vertex.PoseStack;
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
@@ -46,6 +47,12 @@ public class PGeckoModelLoader implements PModelLoader
 	public ResourceLocation id()
 	{
 		return ID;
+	}
+
+	@Override
+	public void applyItemTransform(PoseStack poseStack)
+	{
+		poseStack.translate(0.5f, 0.51f, 0.5f);
 	}
 	
 	@Override